Toggle Data Layer on Admin Click
When clicking on an admin, the visible data layer toggles between
`Reported HDI` and `Estimated HDI`. When a country is selected, the
`onAdminClick` prop for `Map.jsx` is set to the `onAdminClick` action,
which the map calls on click (the map calls whichever of
`onCountryClick` and `onAdminClick` is not `null`). This action toggles
the visible layer and passes a new state for the `layers` key to the
`onAdminSelectUpdater` reducer, which then updates the state.

The rest of the state is not updated, which is why the sidebar is not
changed. This may also be why the data info box and admin outlines do
not show after clicking.